refactor: split matrix event cache lifecycle and storage
